+++ packages/moonshotai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/moonshotai
 
+## 2.0.3
+
+### Patch Changes
+
+- d999bdf: fix (provider/moonshotai): include usage when streaming
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - @ai-sdk/provider@3.0.8
+  - @ai-sdk/openai-compatible@2.0.28
+  - @ai-sdk/provider-utils@4.0.14
+
 ## 2.0.1
 
 ### Patch Changes